upgrade iced 0.13 → 0.14
Main breaking changes: application boot signature (.run_with → .run with
state-returning first arg, title via builder), markdown::Url → Uri and
view() drops style arg (now in Settings), text_editor Content::cursor_position
→ cursor() returning Cursor struct, scrollable::Id → widget::Id with
scroll_to moved to widget::operation, canvas::Program::update returns
Option<Action<Message>> with &Event, horizontal_rule helper removed (use
rule::horizontal), pane_grid::on_resize/spacing want Pixels (no u16
auto-conversion), iced::stream::channel takes AsyncFnOnce directly.

clippy + tests green; app launches and auto-opens project successfully.

+++ crates/letswrite-app/src/app.rs
@@ -275,8 +275,8 @@ impl App {
             };
             pane_grid::Content::new(body)
         })
-        .on_resize(SPLITTER_GRAB_LEEWAY, Message::PaneResized)
-        .spacing(SPLITTER_THICKNESS)
+        .on_resize(f32::from(SPLITTER_GRAB_LEEWAY), Message::PaneResized)
+        .spacing(f32::from(SPLITTER_THICKNESS))
         .style(splitter_highlight_style);
 
         container(pane_grid)
